requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
google-genai
python-multipart
